+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### X.X.X (In progress)
 **SDK**
-- **Feature:** Added `MiniApp.getManifest` interface to retrieve the meta-data of a MiniApp.
+- **Feature:** Added `MiniApp.getMiniAppManifest` interface to retrieve the meta-data of a MiniApp.
 
 **Sample App**
 - **Feature:** Added first-time launching screen before downloading any miniapp.